geaflow/geaflow-infer/src/main/resources/infer/inferRuntime/requirements_paddle.txt
# PaddlePaddle requirements for GeaFlow-Infer PaddleSpatial support
#
# Licensed to the Apache Software Foundation (ASF) under one
# or more contributor license agreements.  See the NOTICE file
# distributed with this work for additional information
# regarding copyright ownership.  The ASF licenses this file
# to you under the Apache License, Version 2.0 (the
# "License"); you may not use this file except in compliance
# with the License.  You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing,
# software distributed under the License is distributed on an
# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
# KIND, either express or implied.  See the License for the
# specific language governing permissions and limitations
# under the License.
#
# NOTE: PaddlePaddle itself is NOT listed here because the GPU/CPU wheel selection
# depends on the target CUDA version and is handled by install-infer-env.sh.
# The script will install paddlepaddle or paddlepaddle-gpu before running pip on
# this file.
#
# Version pins should be updated whenever the production cluster's PaddlePaddle
# version is upgraded. Use exact version matches (==) for reproducible builds.

# Paddle Graph Learning – GNN primitives used by PaddleSpatial models
pgl>=2.2.4

# PaddleSpatial – spatial graph models (SA-GNN, GeomGCN, etc.)
paddlespatial>=0.1.0

# Numerical computing
numpy>=1.21.0,<2.0.0

# Scientific computing (used by PaddleSpatial data utilities)
scipy>=1.7.0

# Process monitoring (used by the infer-env health-check)
psutil>=5.9.0
